feat: common notifications types to shared
