+++ src/observatory/src/lib.rs
@@ -16,9 +16,12 @@ use crate::http::response::transform_response;
 use crate::memory::{get_memory_upgrades, init_runtime_state, init_stable_state, STATE};
 use crate::notify::store_and_defer_notification;
 use crate::random::defer_init_random_seed;
-use crate::store::heap::{delete_controllers, get_controllers, set_controllers as set_controllers_store, set_env as set_env_store};
+use crate::store::heap::{
+    delete_controllers, get_controllers, set_controllers as set_controllers_store,
+    set_env as set_env_store,
+};
 use crate::store::stable::get_notifications;
-use crate::types::interface::{GetNotifications, NotifyArgs, NotifyStatus};
+use crate::types::interface::{GetNotifications, NotifyStatus};
 use crate::types::state::{Env, HeapState, State};
 use crate::upgrade::types::upgrade::UpgradeStableState;
 use ciborium::into_writer;
@@ -26,7 +29,7 @@ use ic_cdk::api::management_canister::http_request::{HttpResponse, TransformArgs
 use ic_cdk::{caller, storage, trap};
 use ic_cdk_macros::{export_candid, init, post_upgrade, pre_upgrade, query, update};
 use junobuild_shared::controllers::init_controllers;
-use junobuild_shared::types::interface::{DeleteControllersArgs, SetControllersArgs};
+use junobuild_shared::types::interface::{DeleteControllersArgs, NotifyArgs, SetControllersArgs};
 use junobuild_shared::types::state::Controllers;
 use junobuild_shared::upgrade::write_pre_upgrade;
