+++ src/observatory/src/notify.rs
@@ -1,9 +1,9 @@
 use crate::http::send::send_notification;
 use crate::store::stable::insert_notification;
-use crate::types::interface::NotifyArgs;
 use crate::types::state::Notification;
 use ic_cdk::{spawn, trap};
 use ic_cdk_timers::set_timer;
+use junobuild_shared::types::interface::NotifyArgs;
 use std::time::Duration;
 
 pub fn store_and_defer_notification(notify_args: &NotifyArgs) {
